整理尽量使用UBOOT的代码配置

+++ bsp/imx6ul/include/asm/posix_types.h
@@ -19,9 +19,9 @@
  * assume GCC is being used.
  */
 
-typedef unsigned short		__kernel_dev_t;
-typedef unsigned long		__kernel_ino_t;
-typedef unsigned short		__kernel_mode_t;
+typedef short		__kernel_dev_t;
+typedef unsigned short		__kernel_ino_t;
+typedef unsigned int		__kernel_mode_t;
 typedef unsigned short		__kernel_nlink_t;
 typedef long			__kernel_off_t;
 typedef int			__kernel_pid_t;
@@ -41,8 +41,8 @@ typedef int			__kernel_ptrdiff_t;
 
 typedef long			__kernel_time_t;
 typedef long			__kernel_suseconds_t;
-typedef long			__kernel_clock_t;
-typedef int			__kernel_daddr_t;
+typedef unsigned long			__kernel_clock_t;
+typedef long			__kernel_daddr_t;
 typedef char *			__kernel_caddr_t;
 typedef unsigned short		__kernel_uid16_t;
 typedef unsigned short		__kernel_gid16_t;
