+++ bsp/imx6ul/include/asm/types.h
@@ -14,8 +14,8 @@ typedef unsigned char __u8;
 typedef __signed__ short __s16;
 typedef unsigned short __u16;
 
-typedef __signed__ int __s32;
-typedef unsigned int __u32;
+typedef __signed__ long __s32;
+typedef unsigned long __u32;
 
 #if defined(__GNUC__)
 __extension__ typedef __signed__ long long __s64;
